LayoutTests/webgl/2.0.0/conformance2/glsl3/uniform-location-length-limits-expected.txt

This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/glsl3/uniform-location-length-limits.html
uniformLoc should be non-null. Was null
getError expected: NO_ERROR. Was INVALID_VALUE :
2 failures reported
6 passes reported
FAIL
